fix: Fix reinitializing with mobx classes & computed. (#31)
The new approach of "keep the old form state on init changes"
doesn't work well for mobx classes, b/c the identity of the mobx
proxy is much more important, and the caller generally assumes
that the (latest) mobx instance will always be the source of truth.

So, if we see that we're being initialized with a class (i.e. mobx
proxy), opt back into the "create new form state each time the
init data changes".