feat: Add useFormStates.readOnly, improve object/field interaction. (#60)
* feat: Add useFormStates.readOnly, improve object/field interaction.

I haven't been entirely happy about how the interaction in the past
of having `formState.readOnly = ...` set and then also adding in
per-field `formState.firstName.readOnly = ...` settings.

The current version was kind of ~last-write-win-ish, i.e. you could
set the base `formState` into `readOnly` but then override it with
a field-level set.

This wasn't terrible, but it always wasn't very declarative.

This PR changes to the approach that each field-level read-only-ness
is now just a combination of:

    formState.readOnly || fieldState.readOnly

I.e. either form-level or field-level being true will lock the field.

This technically removes the ability, which the previous setup had,
of allowing a form-level readOnly=true but allowing a specific field
to opt-out and go back to readOnly=false.

However, I like the new "form || field" semantics enough that I think
it's fine to give up this capability, which was kind of made up anyway.
If a page really needed to have "some fields opt _out_ of being read
only", it should probably just not be setting the form-level read only
anyway, and just directly control every form field.

I'm not worried about this breaking change because in Blueprint we hardly
ever use this formState-level + field-level readOnly pattern, and instead
just override the readOnly props on the way into our field components.

But really I'm hoping we can do less of that "override the field
component prop" and do more "set the field state whether it is/is not
read only", with this ergonomic improvement in how the two form-level /
field-level read only flags interact.

* Force a boolean.