feat: Add init.onlyOnce. (#111)
A previous refactoring removed the ability to do `init: pojo` in
the API, and instead consolidated usage on `init: { input: pojo }`.

This was good, but a surprising artifact of `init: pojo` was that it did
not re-init as the identity of pojo changes (vs. the `init.input = pojo`
whcih *does* re-init as the identity changes).

This dictomoy between "one re-inits and one does not" was fairly
aribtrary and was a big reason behind removing the "two ways of doing
things".

That said, in rolling out the change to internal-frontend, a number of
callers were enjoying the implicit useMemo-ification of their `init:
pojo`, and so this PR adds an `init.onlyOnce` flag that basically
restores the "don't watch the init.input identity" behavior.