feat: Avoid warnings when using useFormStates.
Currently when tables call `getFormState(fragment)`, from within their
render method (mobx observer), we immediately diff the fragment against
the prior fragment, and if it's different, push the update into the
formState proxy.

This "update a proxy in a render method" will tick/dirty any other
components/observers that are also watching this mobx proxy.

The regular useFormState avoids this by writing to the proxy via a
useEffect, which moves the dirty-via-setState out of render methods.

This PR is a spike/proof-of-concept to apply the same useEffect
approach to useFormStates, by turning getFormState effectively into
a hook itself.